backend/services/service-results/requirements.txt
quart
motor
python-dotenv
pydantic
pydash
pytest
pytest-asyncio
pytest-mock